requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0
python-dotenv>=1.0.0
streamlit>=1.32.0
pytest>=7.0.0
pytest-cov>=4.0.0
typing-extensions>=4.5.0
numpy>=1.24.0
pandas>=2.0.0
